Split compat files into go 1.8 and pre-go 1.8 files
